requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai>=1.0.0
python-docx
pdfplumber